requirements.txt
black==25.1.0
Django==4.2.24
django-cors-headers==3.11.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
numpy==1.22.2
pandas==1.3.4
pillow==11.3.0
psycopg2-binary==2.9.10
requests==2.28.2
selenium==4.1.0
